isole les fonctions d'écritures
