+++ src/grand-livre.js
@@ -1,5 +1,5 @@
 import { trouverCompte, formatToCurrency } from './utils.js';
-import { lignesEnEcritures } from './ecritures.js';
+import { lignesEnEcritures } from './gestion-ecritures.js';
 
 /**
  * Crée un grand livre comptable à partir des données JSON et de l'année courante.
