+++ src/journal.js
@@ -1,5 +1,5 @@
 import { formatToCurrency } from './utils.js';
-import { lignesEnEcritures } from './ecritures.js';
+import { lignesEnEcritures } from './gestion-ecritures.js';
 
 /**
  * Crée un journal comptable à partir des données JSON et de l'année courante.
